requirements.txt
PyYAML>=6.0.0,<7.0.0
python-dotenv==1.2.1
